Extract the files view to public/views/files.js
Verbatim move of app.js section 11 (file tree, creation menus, editor
surfaces, wikilink navigation) plus the section-1 file-surface lifecycle
(module loaders, Tiptap init/teardown, viewers and artifact-review mounts,
the external-edit guard, closeOpenFile) into a UMD view module,
RundockFilesView, republished on the root object for the inline handlers,
the generated onclick wikilinks, and cross-section callers.

Shared state stays in app.js, applied per identifier: every declaration
with a retained reader or writer (WS dispatch, find bar, init listeners,
routing, palette) keeps its home; moved functions reach it through the
global lexical environment at call time. TREE_ICONS and CREATABLE_TYPES
stay because their declarations read FilesMenuModel at load time, which
the side-effect-free factory cannot do without breaking clean Node
require. The two document-level menu-close listeners stay as top-level
wiring. The four dynamic import() specifiers become absolute because
import() resolves against the importing script's URL; they load the same
URLs as before. All 44 function bodies are otherwise byte-identical to
the originals at column 0. app.js 4,921 -> 4,154 lines.
