+++ public/index.html
@@ -1387,6 +1387,7 @@ mark.find-match.current,
 <script src="/views/settings.js"></script>
 <script src="/views/team.js"></script>
 <script src="/views/profile.js"></script>
+<script src="/views/files.js"></script>
 <script src="/app.js"></script>
 </body>
 </html>
